MVP self-host
