+++ docker-compose.yml
@@ -1,20 +1,20 @@
 version: "3.8"
 
 services:
+  kv:
+    container_name: express_db
+    image: eqalpha/keydb
+    entrypoint: ["keydb-server", "/etc/keydb/keydb.conf"]
+    expose:
+      - 6379
+    volumes:
+      - "./docker/keydb.conf:/etc/keydb/keydb.conf"
+
   app:
     container_name: express_service
     build:
       context: ./
       dockerfile: ./docker/Dockerfile
-    image: ubuntu:22.04
-    environment:
-      - SELF_HOST=1
-      - BUN_CANARY=1
-      - NODE_ENV=production
-    entrypoint: ["/web/bun", "index.js"]
-    ulimits:
-      memlock:
-        soft: -1
-        hard: -1
+    image: denoland/deno:alpine
     ports:
-      - "${API_HOST:-0.0.0.0}:${API_PORT:-3000}:3000/tcp"
+      - "${API_HOST:-0.0.0.0}:${API_PORT:-3000}:8000/tcp"
