+++ docker/keydb.conf
@@ -1,0 +1,11 @@
+bind 0.0.0.0
+port 6379
+
+daemonize no
+dbfilename express_db.rdb
+
+# Save every 30 seconds if at least 1 key has changed
+save 30 1
+
+# Save every 5 seconds if at least 25 keys have changed
+save 5 25
