--- docker/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "GmodExpress",
-  "private": true,
-  "version": "1.0.0",
-  "description": "GmodExpress is a web application that lets you send data to, and receive data from your clients out-of-band",
-  "main": "index.js",
-  "license": "ISC",
-  "type": "module",
-  "dependencies": {
-    "hono": "^2.4.1"
-  }
-}
