+++ index.js
@@ -1,84 +1,15 @@
-import { Hono } from "hono"
-
-const app = new Hono()
 const expiration = 60 * 60 * 24
 const maxDataSize = 24 * 1024 * 1024
 
-const isSelfHost = process.env.SELF_HOST === "1"
-console.log(process.env.SELF_HOST, isSelfHost)
-
-if (isSelfHost === true) {
-  console.log("Is self host, setting env")
-  // If self hosting, use our KeyDB in place of KV
-
-  const Database = require("bun:sqlite").Database
-  const db = new Database(":memory:")
-  db.run(
-    "CREATE TABLE express (id INTEGER PRIMARY KEY AUTOINCREMENT, key TEXT, value BLOB, expiration INTEGER)"
-  )
-
-  const expressStub = {
-    get: (key, metadata) => {
-      return new Promise((resolve, reject) => {
-        console.log("(db) GET", key)
-        const row = db.query("SELECT value FROM express WHERE key = $key").get({ $key: key })
-        console.log(row)
-
-        const isBuffer = metadata && (metadata.type === "arrayBuffer")
-
-        const val = row.value
-        if (isBuffer) {
-          console.log("Returning buffer", key)
-          resolve(Buffer.from(val, "base64"))
-        } else {
-          console.log("Returning string", key)
-          resolve(val)
-        }
-      })
-    },
-
-    put: (key, value, metadata) => {
-      console.log("(db) PUT", key, metadata)
-      let ttl = expiration
-      let isbuffer = false
-      if (metadata) {
-        if (metadata.expirationTtl) {
-          ttl = metadata.expirationTtl
-        }
-        if (metadata.type === "arrayBuffer") {
-          isbuffer = true
-        }
-      }
-      return new Promise((resolve, reject) => {
-        const ex = Date.now() + ttl
-        if (isbuffer === true) {
-          value = Buffer.from(value)
-        }
-        db.run( "INSERT INTO express (key, value, expiration) VALUES ($key, $value, $expiration)", { $key: key, $value: value, $expiration: ex })
-        resolve()
-      })
-    }
-  }
-
-  // setInterval(() => {
-  //   db.run("DELETE FROM express WHERE expiration < ?", Date.now())
-  // }, 1000 * 10 )
-
-  app.use("*", async (c, next) => {
-    c.env.GmodExpress = expressStub
-
-    await next()
-  })
-} else {
-  console.log("Is not self host, using KV")
-}
+const isSelfHost = !!Deno
+const { app, serve } = await import(isSelfHost ? "./selfhost.js" : "./cloudflare.js")
 
 const makeMetadata = (c, extraMetadata) => {
   return {
     expirationTtl: expiration,
     metadata: {
       ...extraMetadata,
-      remote: c.req.header("CF-Connecting-IP")
+      remote: c.req.header("CF-Connecting-IP") || c.req.header("X-Forwarded-For")
     }
   }
 }
@@ -86,17 +17,12 @@ const makeMetadata = (c, extraMetadata) => {
 async function validateRequest(c, token) {
   // TODO: Do a sanity check on expiration time too
   const expected = await c.env.GmodExpress.get(`token:${token}`)
-  console.log("Allowed", !!expected)
-
   return !!expected
 }
 
 async function putData(c, data) {
   const id = crypto.randomUUID()
   const metadata = makeMetadata(c)
-
-  console.log("PUT", `size:${id}`, data.byteLength)
-  console.log("PUT", `data:${id}`)
 
   await Promise.all([
     c.env.GmodExpress.put(`size:${id}`, data.byteLength, metadata),
@@ -108,17 +34,14 @@ async function putData(c, data) {
 
 async function putToken(c, token) {
   const now = Date.now().toString()
-  console.log("PUT", `token:${token}`, now)
   await c.env.GmodExpress.put(`token:${token}`, now, makeMetadata(c))
 }
 
 async function getData(c, id) {
-  console.log("GET", `data:${id}`)
   return await c.env.GmodExpress.get(`data:${id}`, { type: "arrayBuffer" })
 }
 
 async function getSize(c, id) {
-  console.log("GET", `size:${id}`)
   return await c.env.GmodExpress.get(`size:${id}`)
 }
 
@@ -143,7 +66,6 @@ async function readRequest(c) {
 
   const id = c.req.param("id")
   const data = await getData(c, id)
-  console.log("Retrieved data, preparing to send", data.byteLength)
   return c.body(data, 200, { "Content-Type": "application/octet-stream" })
 }
 
@@ -155,7 +77,6 @@ async function readSizeRequest(c) {
   }
 
   const id = c.req.param("id")
-  console.log("Getting size for", id)
   return c.json({size: await getSize(c, id)})
 }
 
@@ -166,10 +87,7 @@ async function writeRequest(c) {
     return c.text("", 401)
   }
 
-  console.log(c.req)
   const data = await c.req.arrayBuffer()
-  console.log(data)
-  console.log("Got data", data.byteLength)
   if (data.byteLength > maxDataSize) {
     return c.text("Data exceeds maximum size of " + maxDataSize, 413)
   }
@@ -191,6 +109,5 @@ app.get("/v1/revision", async (c) => {
 
 app.get("*", async (c) => c.text("Not Found - you may need to update the gm_express addon!", 406))
 
-const exported = isSelfHost ? { port: 3000, fetch: app.fetch } : app
-console.log(exported)
-export default (isSelfHost ? { port: 3000, fetch: app.fetch } : app)
+export default app
+serve(app.fetch)
