+++ workers.js
@@ -1,0 +1,6 @@
+import { Hono } from "hono"
+
+const app = new Hono()
+const serve = () => {}
+
+export { app, serve }
